refactor: 도메인명 login -> user로 변경
