--- src/main/java/com/codiapp/codi/domain/login/dto/response/LoginResponseDTO.java
@@ -1,4 +1,0 @@
-package com.codiapp.codi.domain.login.dto.response;
-
-public record LoginResponseDTO (Long id, String username, String email) {
-}
