+++ src/main/java/com/codiapp/codi/domain/user/dto/request/LoginRequestDTO.java
@@ -1,4 +1,4 @@
-package com.codiapp.codi.domain.login.dto.request;
+package com.codiapp.codi.domain.user.dto.request;
 
 public record LoginRequestDTO (String email, String password) {
 }
