+++ src/main/java/com/codiapp/codi/domain/user/dto/response/LoginResponseDTO.java
@@ -1,0 +1,12 @@
+package com.codiapp.codi.domain.user.dto.response;
+
+import com.codiapp.codi.domain.user.entity.UserRole;
+import lombok.Builder;
+
+@Builder
+public record LoginResponseDTO (
+        Long id,
+        String username,
+        String email,
+        UserRole role
+) {}
