+++ src/main/java/com/codiapp/codi/domain/user/repository/UserRepository.java
@@ -1,12 +1,12 @@
-package com.codiapp.codi.domain.login.repository;
+package com.codiapp.codi.domain.user.repository;
 
 import java.util.Optional;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.codiapp.codi.domain.login.entity.User;
+import com.codiapp.codi.domain.user.entity.User;
 
 public interface UserRepository extends JpaRepository<User, Long>{
-    Optional<User> findByEmail(String email);    
+    Optional<User> findByEmail(String email);
 }
